Update 4
Redesigned the catalog page to be a grid view.
Added a top and side bar to the catalog page. Side bar contains shelves to filter the catalog of books into such as all books, favorites and read. Top bar includes the new location for the search bar (functionality not yet added).
Added the ability to favorite books.
Reorganized App.jsx and added some comments for organizing.
Duplicate prevention was added to the search page. The add button shows "Already added" if the book is already in the catalog.